requirements.txt
pipi
numpy
requests 
datetime
beautifulsoup4
progressbar
IPython
django-bootstrap4
